Introduce Multilevel And UltraSound Matrices
1. We add a nice "Meta"-Class, which makes use of the already
existing classes Circulant, Fourier, Kron, Product and Partial.
It generalizes convolutions to multiple dimensions and also
includes efficient zeropadding for maximal performance.

2. Add MLToeplitz and finalize MLCirculant

3. Add Block Diagonal Matrix, which gets used to build up a
dictionary matrix used in ultrasonic nondestructive testing

4. Add _getNormalized Functions To Convolutional Matrices
We add specialized and very fast algorithms to return a normalized
version of the single- and multilevel convolutionary transforms
as well as the UltraSound class. This dramatically shortens the
initialization time of OMP and other Algorithms demanding a
column normalized transform.

5. Fix docu title in Permutation