Fix {Circulant,Toeplitz}.normalized accuracy
The classes
  - Circulant
  - MLCirculant
  - MLToeplitz
  - MLUltraSound
  - Toeplitz
showed a reduced accuracy of only single precision for their normalized
versions, which were generated by accessing the normalized property of the
aforementioned classes. The fix consisted of adding type consistency checks
for the internal storage arrays of each class, such that their type matches
the type of the matrix itself.

These checks are neccessary as real-valued definition arrays will still
generate a complex-valued matrix due to the cast-to-complex behaviour of
the underlying Fourier class.

If more cunning handling of the Fourier is added to fastmat, this issue
should resolve and the casts should become obsolete.