Introduce input sanity checks to inspect.test
If an empty input array was passed to a transform test, the test always
succeeded since transforming a zero vector also produces a zero vector.
This was fixed by rejecting input vectors with ||arrX|| == 0.

Also, if a transform changed the given input vector accidentially this
went unnoticed before. Now array checks of arrInput in testForward(),
testBackward() and testAlgorithm() are in place to precent this.