Modify LFSRCirculant to accept full polynomial
Previously, an error was thrown when a tap position was specified that
exceeded the size of the register. As the tap configuration directly
corresponds to the generating polynomial -- and this always has a x^k
factor with k being the order of the polynomial (size of the register) --
it is sensible to also accept the full notation of the polynomial.