Fix bug in Matrix.largestSingularV{alue,ectors}
Due to a forgotten return the properties returned the proper value only for
the first invokation (when getLargest... returned the value straight away)
or never in the case of direct overwriting of ._largest...

This is resolved by adding the missing returns. This can not really be
tested automatically as the checks would need to be garnered specifically
for every property to be applied upon.