Optimize Fourier Bluestein initializaton. Resolves #45
The two calls of `np.exp(...)` in the Bluestein section of the Fourier
class `.__init__()` are now replaced by explicit cos/sin computations
on empty arrays. Initialization with zeros of `vecConv` is reduced to
only those sections that are not overwritten by any coefficients anyway.
The trigonometric functions are used such that symmetries are exploitet,
effectively reusing each single operation twice, resulting in a two- to
six-fold increase over the original `np.exp(...)` calls (depending on
the linear algebra system that is compiled against).