Fix Fourier Bluestein problem. Solves #115
The regression occurred in #d22813f0, where the np.exp() replacement
(for speed enhancement purposes) did not consider real and imaginary
part properly.
Unfortunately, the test case for the activated-Bluestein-case did not
reliably trigger Bluestein expansion, which is why it did not show up
earlier.

Both issues are addressed by this commit, resolving the issue.